requirements.txt
# AI APIs
anthropic>=0.34.0
openai>=1.0.0
google-genai>=1.0.0

# Collaborative Filtering
scikit-surprise>=1.1.3
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# TMDB API
requests>=2.31.0

# Web Backend
flask>=3.0.0
flask-cors>=4.0.0

# Caching
diskcache>=5.6.0

# Utilities
python-dotenv>=1.0.0

# Web Scraping (for Letterboxd)
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
